fix: Dodatni MISRA C in ISO 26262 compliance fixes (5 issues)
CRITICAL: Compilation Error
- ltc_scanner.c: Dodana manjkajoča konstanta LTC_EVENT_CONFIG_ERROR (0x06U)
- Uporabljena v ltc_sm_read_group bounds check (linija 449)
- Rešuje compilation error: undefined reference

Type Conversion Warnings (MISRA 10.1)
- bcu_scheduler.c:281: Explicit cast uint64_t → uint32_t z komentarjem
- app_main.c:176: Komentar za intentional 32-bit truncation
- Intentional truncation for 32-bit timing calculations dokumentiran

HIGH #20: Volatile Qualifier
- app_main.c: Odstranjen nepotrebni volatile iz task timestamps
- fast_task_last_ms, medium_task_last_ms, slow_task_last_ms
- Ne uporabljeno v ISR kontekstu, volatile ni potreben
- MISRA compliance - volatile samo za ISR-accessed variables

HIGH #14: Error_Handler Interrupt Management
- main.c:634: Odstranjen __disable_irq() iz Error_Handler
- Omogoča watchdog ISR za system reset
- ISO 26262 ASIL-B: Interrupts enabled for recovery mechanism
- Watchdog lahko resetira sistem tudi med error handlingom

Compliance Status:
- MISRA C:2012: 10.1 (type conversion), volatile usage
- ISO 26262 ASIL-B: Error recovery, interrupt management
- 1 CRITICAL + 4 issues resolved
- Total resolved: 8 CRITICAL + 8 HIGH = 16 violations

+++ Application/Src/bcu_scheduler.c
@@ -278,7 +278,8 @@ uint32_t Scheduler_WCET_Start(const char *task_name)
 Status_t Scheduler_WCET_Stop(const char *task_name, uint32_t start_us, uint32_t budget_us)
 {
     Status_t status = STATUS_OK;
-    uint32_t end_us = Timestamp_GetMicros();
+    /* MISRA 10.1: Explicit cast - intentional truncation for 32-bit timing */
+    uint32_t end_us = (uint32_t)Timestamp_GetMicros();
     uint32_t elapsed_us = end_us - start_us;
 
     /* Find or create WCET stats entry */
